refactor: state atom file name
